Combine the query object and the presenter.
